Refactor: Divide statistics table component
통계 테이블을 세개의 컴포넌트로 분리
beta의 경우 테이블이 깨져서 라우트에서 임시로 div로 감싸주었음
