+++ src/routes/Domestic/DoCorrelation.jsx
@@ -1,15 +1,21 @@
 import React from "react";
-import CorrelationTable from "@components/CorrelationTable";
+import TableHeader from "@components/Table/TableHeader"
+import CorrelationTable from "@components/Table/CorrelationTable";
 import CorrelationChart from "@components/CorrelationChart"
 import {domesticSample} from "@utils/statisticsData"
+import {CorrTableWrapper} from "./style"
 
 const DoCorrelation = () => {
-  return (
-      <>
-        <CorrelationTable data={domesticSample} />
-        <CorrelationChart />
-      </>
-  );
+  return (<>
+    <CorrTableWrapper>
+      <TableHeader 
+        data={domesticSample.corr} 
+        title={domesticSample.title[1]} 
+      />
+      <CorrelationTable data={domesticSample} />
+    </CorrTableWrapper>
+    <CorrelationChart />
+  </>);
 };
 
 export default DoCorrelation;
