+++ src/routes/Domestic/style.js
@@ -7,3 +7,11 @@ export const DomesticWrapper = styled.div`
     margin-top: 20px;
   }
 `;
+
+export const CorrTableWrapper = styled.div`
+  float: left;
+  width: 23.913043%;
+  height: calc(100vh - 265px);
+  border-radius: 0 0 12px 12px;
+`;
+
