+++ src/routes/Global/GoBeta.jsx
@@ -1,10 +1,12 @@
 import React from "react";
-import BetaTable from "@components/BetaTable";
+import BetaTable from "@components/Table/BetaTable";
+import TableHeader from "@components/Table/TableHeader"
 import {globalSample} from "@utils/statisticsData"
 
 const GoBeta = () => {
   return (
     <>
+      <TableHeader data={globalSample.beta} title={globalSample.title} /> 
       <BetaTable data={globalSample} />
     </>
   );
