+++ src/routes/Global/GoCorrelation.jsx
@@ -1,11 +1,21 @@
 import React from "react";
-import CorrelationTable from "@components/CorrelationTable";
+import TableHeader from "@components/Table/TableHeader"
+import CorrelationTable from "@components/Table/CorrelationTable";
+import CorrelationChart from "@components/CorrelationChart"
 import {globalSample} from "@utils/statisticsData"
+import {CorrTableWrapper} from "../Domestic/style"
 
 const GoCorrelation = () => {
   return (
     <>
-      <CorrelationTable data={globalSample} />
+      <CorrTableWrapper>
+        <TableHeader 
+          data={globalSample.corr} 
+          title={globalSample.title[1]} 
+        />
+        <CorrelationTable data={globalSample} />
+      </CorrTableWrapper>
+      <CorrelationChart />
     </>
   );};
